#1287 rename file containing DataTypeSymbol
