+++ src/psyclone/psyir/symbols/data_type_symbol.py
