+++ src/psyclone/tests/psyir/symbols/data_type_symbol_test.py
